libbjxa: Turn the header parsing upside down
Despite the lack of common sense in ISO C99 and POSIX.1 (and yes that
includes POSIX.1-2017 released this summer) it's actually possible to
make a single read and then parse: treat the memory buffer as tape.

That doesn't change the semantics of the functions but it's a lot
cleaner... In this form it's also easier to assert that a complete
header is read and manual testing showed that the loop field was
missing.