test: Basic coverage of XA decoding
It pains me to check in binary files but they aren't too big and
shouldn't need to change over time. Obviously the checksums were
generated by the library in its current state so it basically
agrees with itself at this point, but the generated wave files
have been manually checked with human ears. In fact, the library
was tested with 4485 XA files found in the wild (3700 unique files
once deduplicated) including long-running tracks (over a minute)
and everything sounded right.

A simple test suite like that is enough to get to 98% coverage of
the library!