build: Add AM_CONDITIONAL to --with and --enable
This simplifies how optional features are handled and removes some
duplication around the option conditions. The condition is even hidden
behind the WITH/ENABLE macros.

And consequently use AM_IF when the same condition needs to be checked
at configure time in addition to make time.

To fit the template lcov's conditional had to be renamed.