libbjxa: Single-profile encoding
The encoder does not try to find the best profile for a given block but
uses the most neutral profile for now. It has no adaptive gain factors
and can't scale samples. It should be equivalent to raw compression from
16 PCM samples to 4, 6 or 8 bits, simply keeping the N most significant
bits.

The conversion is currently as simple as:

    bjxa_sample = (uint16_t)wav_sample >> (16 - N);

The right shift does not happen during the conversion though, it is the
deflate callback that in effect loses the least significant bits.